+++ app/(tabs)/_layout.tsx
@@ -31,7 +31,7 @@ export default function TabLayout() {
 
   return (
     <Tabs
-      tabBar={(props) => <TabBar {...props} />}
+      // tabBar={(props) => <TabBar {...props} />}
       screenOptions={{
         headerShown: true,
         headerTitle: '',
@@ -46,17 +46,45 @@ export default function TabLayout() {
           );
         },
       }}>
-      <Tabs.Screen name="index" options={{title: 'Stron główna'}} />
+      <Tabs.Screen
+        name="index"
+        options={{
+          tabBarIcon: ({color}) => <TabBarIcon name="home" color={color} />,
+          title: 'Stron główna',
+        }}
+      />
       <Tabs.Screen
         name="records"
-        options={{title: 'Wydatki/Wpływy', headerShown: true}}
+        options={{
+          tabBarIcon: ({color}) => <TabBarIcon name="cash" color={color} />,
+          title: 'Wydatki/Wpływy',
+          headerShown: true,
+        }}
       />
-      <Tabs.Screen name="addnew" options={{title: 'Dodaj'}} />
+      <Tabs.Screen
+        name="addnew"
+        options={{
+          tabBarIcon: ({color}) => <TabBarIcon name="add" color={color} />,
+          title: 'Dodaj',
+        }}
+      />
       <Tabs.Screen
         name="summary"
-        options={{title: 'Podsumowanie', headerShown: true}}
+        options={{
+          tabBarIcon: ({color}) => (
+            <TabBarIcon name="bar-chart" color={color} />
+          ),
+          title: 'Podsumowanie',
+          headerShown: true,
+        }}
       />
-      <Tabs.Screen name="settings" options={{title: 'Ustawienia'}} />
+      <Tabs.Screen
+        name="settings"
+        options={{
+          tabBarIcon: ({color}) => <TabBarIcon name="settings" color={color} />,
+          title: 'Ustawienia',
+        }}
+      />
     </Tabs>
   );
 }
