+++ app/(tabs)/addnew/_layout.tsx
@@ -3,7 +3,7 @@ import {Stack} from 'expo-router';
 const StackLayout = () => {
   return (
     <Stack screenOptions={{headerShown: false}}>
-      <Stack.Screen name="index" options={{title: 'Dodaj Wydatek'}} />
+      <Stack.Screen name="expense" options={{title: 'Dodaj Wydatek'}} />
       <Stack.Screen name="income" options={{title: 'Dodaj Wpływy'}} />
     </Stack>
   );
